Fix vision_rag cohere .float field and deepseek agno 2.x migration
- vision_rag/vision_rag.py: cohere's embeddings field is float_ (alias 'float'
  only affects serialization); .float raises AttributeError, swallowed into
  'Failed to get embedding', so the app never works. Use .float_ (4 sites).
- deepseek_local_rag_agent: import OllamaEmbedder from agno.knowledge.embedder
  (agno 2.x moved it; old path ModuleNotFoundErrors), replace show_tool_calls
  with debug_mode (removed from Agent in 2.x), pin agno>=2.2.10, add pypdf
  (PyPDFLoader is used). Mirrors the already-migrated qwen_local_rag sibling.

Fixes #1023

+++ rag_tutorials/vision_rag/vision_rag.py
@@ -135,8 +135,8 @@ def compute_image_embedding(base64_img: str, _cohere_client) -> np.ndarray | Non
             images=[base64_img],
         )
         
-        if api_response.embeddings and api_response.embeddings.float:
-            return np.asarray(api_response.embeddings.float[0])
+        if api_response.embeddings and api_response.embeddings.float_:
+            return np.asarray(api_response.embeddings.float_[0])
         else:
             st.warning("Could not get embedding. API response might be empty.")
             return None
@@ -320,11 +320,11 @@ def search(question: str, co_client: cohere.Client, embeddings: np.ndarray, imag
             texts=[question],
         )
 
-        if not api_response.embeddings or not api_response.embeddings.float:
+        if not api_response.embeddings or not api_response.embeddings.float_:
             st.error("Failed to get query embedding.")
             return None
 
-        query_emb = np.asarray(api_response.embeddings.float[0])
+        query_emb = np.asarray(api_response.embeddings.float_[0])
 
         # Ensure query embedding has the correct shape for dot product
         if query_emb.shape[0] != embeddings.shape[1]:
